UI: Add UIButtonRow
